+++ braindecode/models/__init__.py
@@ -33,7 +33,7 @@ from .eegnex import EEGNeX
 from .tsinception import TSceptionV1
 from .eegtcnet import EEGTCNet
 from .syncnet import SyncNet
-from .sinc_shallownet import SincShallowNet
+from .sinc_shallow import SincShallowNet
 
 # Call this last in order to make sure the dataset list is populated with
 # the models imported in this file.