+++ docs/whats_new.rst
@@ -53,7 +53,7 @@ Enhancements
 - Adding :class:`braindecode.models.SyncNet` (:gh:`643` by `Bruno Aristimunha`_ )
 - Creating the `FilterBanklayer` module for new models (:gh:`656` by `Bruno Aristimunha`_ )
 - Including PytorchAudio as dependency and remove copied code (:gh:`661` by `Bruno Aristimunha`_)
-- Adding :class:`braindecode.models.SincShallowNet` (:gh:`670` by `Bruno Aristimunha`_ )
+- Adding :class:`braindecode.models.SincShallowNet` (:gh:`671` by `Bruno Aristimunha`_ )
 
 
 Bugs